c128: keyboard matrix + C64 mode with working boot screen
- Wire 11x8 C128 keyboard matrix via CIA1 port callbacks
  - Forward scan (Port B read): columns 0-7 via CIA PA, columns 8-10 via VIC-IIe $D02F
  - Reverse scan (Port A read): columns 0-7 only (8-bit Port A)
  - KeyboardMapper integration with EmuKeySDLMap fallback
- Widen commodore_keyboard_t for C128: MAX_KEYBOARD_COLS 8->16, row_open_contacts uint8_t->uint16_t
  - Fix existing CIA callbacks in C64/VIC-20/C16/PET to cast wider type
- Add C64 BASIC (8KB) and C64 KERNAL (8KB) ROM slots for C64 compatibility mode
  - Correct C64-mode banking: BASIC at $A000-$BFFF, KERNAL at $E000-$FFFF
  - CPU reset on enter_c64_mode() to fetch C64 KERNAL reset vector
- Wire CIA1 interrupt line to IRQ (cursor blink, keyboard scan)
- Fix MOS 8722 MMU CR bit field encoding (bit 0=I/O select, bit 1=BASIC LO)
- Add C128-specific emu_key_t values and keyboard mapper factory

+++ src/systems/commodore/c64/c64_system.cpp
@@ -1542,7 +1542,7 @@ static uint8_t c64_cia1_port_a_read_with_joystick(void* context, uint8_t port_a_
         uint8_t row_select = ~port_b_output;
         for (int row = 0; row < 8; row++) {
             if (row_select & (1 << row)) {
-                col_state &= c64->keyboard->row_open_contacts[row];
+                col_state &= static_cast<uint8_t>(c64->keyboard->row_open_contacts[row]);
             }
         }
     }
